refactor(runtime): move SecretInput into zeroclaw-runtime::cli_input
SecretInput was implemented in the root crate's src/cli_input.rs, which
otherwise re-exports zeroclaw_runtime::cli_input. That made it the only
interactive-input primitive not living beside Input and
TerminalUtf8EraseGuard, and it left the root crate — a facade where 124 of
160 modules are pure re-export shims — owning 226 lines of terminal
implementation.

Move it to zeroclaw-runtime::cli_input with its three unit tests. The root
module becomes the one-line shim the surrounding convention expects;
main.rs needs no change because cli_input::SecretInput still resolves
through it. Availability is unchanged: the root module was already gated on
agent-runtime, and zeroclaw-runtime is only present under that feature.

zeroclaw-runtime takes crossterm with default features only. The root crate
requested the event-stream feature, but SecretInput uses the blocking
event::read() and nothing in the workspace uses EventStream, so the async
event machinery was being compiled for nothing.
